Beginner/file_organiser/pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "file_organiser"
version = "0.1.0"
description = "Help sort your files"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.4.2",
    "python-magic>=0.4.27",
    "rich>=14.2.0",
    "ruff>=0.14.2",
]
